onlinetlabs-backend/pyproject.toml
[tool.poetry]
name = "onlinetlabs-backend"
version = "0.1.0"
description = "OnlineTLabs Backend API"
authors = []
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.128.0"
uvicorn = {extras = ["standard"], version = "^0.40.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.31.0"
alembic = "^1.13.0"
pydantic-settings = "^2.1.0"
bcrypt = "^4.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pydantic = {extras = ["email"], version = "^2.0.0"}
python-dotenv = "^1.2.1"
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
pytest-asyncio = "^1.0.0"
httpx = "^0.28.0"
ruff = "^0.14.0"
aiosqlite = "^0.22.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
